+++ sdk/core/core-util/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-- Added `calculateExponentialDelayInterval` helper that can be used to calculate the next delay interval for exponential delay with jitter. []
+- Added `calculateExponentialDelayInterval` helper that can be used to calculate the next delay interval for exponential delay with jitter. [#30572](https://github.com/Azure/azure-sdk-for-js/pull/30572)
 
 ### Breaking Changes